mjoving helper methods to a util module
