--- services/couchdb/dmdtask/fixtures/21f6c35448f3a37688d482004b0007e8.json
@@ -1,90 +1,0 @@
-{
-  "_id": "c4435f96ed390ebbeeb01354b500054c",
-  "_rev": "10-b21abc3129737c18cb93c5dada779bd5",
-  "format": "marc490",
-  "user": {
-    "name": "Brittny Lapierre",
-    "email": "blapierre@crkn.ca"
-  },
-  "updated": 1629987932.601,
-  "process": {
-    "requestDate": 1629908746.611,
-    "processDate": 1629920071.491,
-    "succeeded": true,
-    "message": ""
-  },
-  "items": [
-    {
-      "message": "bueno",
-      "id": "8_06941",
-      "output": "marc",
-      "label": "I am a bueno label",
-      "parsed": true
-    },
-    {
-      "message": "sehr gut",
-      "id": "8_06941_1",
-      "output": "marc",
-      "label": "I am a sehr gut label",
-      "parsed": true
-    },
-    {
-      "message": "good",
-      "id": "8_06941_2",
-      "output": "marc",
-      "label": "I am a good label",
-      "parsed": true
-    }
-  ],
-  "_attachments": {
-    "2.xml": {
-      "content_type": "text/xml",
-      "revpos": 10,
-      "digest": "md5-Wzec+DXj6BHzdlSVhHaocw==",
-      "length": 497,
-      "stub": true
-    },
-    "1.xml": {
-      "content_type": "text/xml",
-      "revpos": 9,
-      "digest": "md5-NiN52Tr/UuBSxRhJyBbZvg==",
-      "length": 497,
-      "stub": true
-    },
-    "0.xml": {
-      "content_type": "text/xml",
-      "revpos": 8,
-      "digest": "md5-sdKc1TaR2GS3nr9xLZ1MbQ==",
-      "length": 497,
-      "stub": true
-    },
-    "2.json": {
-      "content_type": "application/json",
-      "revpos": 7,
-      "digest": "md5-gC66f+P6yAYgIyiQmNNQuw==",
-      "length": 699,
-      "stub": true
-    },
-    "1.json": {
-      "content_type": "application/json",
-      "revpos": 6,
-      "digest": "md5-gC66f+P6yAYgIyiQmNNQuw==",
-      "length": 699,
-      "stub": true
-    },
-    "0.json": {
-      "content_type": "application/json",
-      "revpos": 5,
-      "digest": "md5-EMfBlnetoUElb212i7XL3A==",
-      "length": 699,
-      "stub": true
-    },
-    "metadata": {
-      "content_type": "application/octet-stream",
-      "revpos": 2,
-      "digest": "md5-TtujNLOPhBUBvOxXo5+cTg==",
-      "length": 5102086,
-      "stub": true
-    }
-  }
-}
